push pronta pra entregar

+++ push_swap/libft/ft_atol.c
@@ -6,17 +6,17 @@
 /*   By: bgomes-b <bgomes-b@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2025/11/03 12:43:57 by bgomes-b          #+#    #+#             */
-/*   Updated: 2026/01/29 21:29:01 by bgomes-b         ###   ########.fr       */
+/*   Updated: 2026/01/31 20:46:24 by bgomes-b         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "libft.h"
 
-int	ft_atol(const char *nptr)
+long	ft_atol(const char *nptr)
 {
-	int	i;
-	int	neg;
-	int	res;
+	long	i;
+	long	neg;
+	long	res;
 
 	i = 0;
 	neg = 1;
